Refactor main.py for flexible path handling and improved error messages
- Enable dynamic file path input for different ECG data types.
- Ensure file paths are only required for 'standard' and 'eco' types.
- Add user-friendly error handling for missing or unnecessary file paths.
- Update argument parsing to conditionally process file names.
